39 duplicate method call refactor

+++ app/models/room.rb
@@ -1,28 +1,25 @@
 class Room < ActiveRecord::Base
   mount_uploader :room_image, MainUploader
+  include Errors
 
+
+  def find_items_for_room
+    return Item.where({"room_id" => self.id})
+  end
+
+  # Method finds all Item Object instances where "room_id" matches the current Room Instance ID && stores the item id in @item_id_array instance variable.
+  # SETS Array of Item Collection IDs to Instance Variable @item_id_array
+  def find_items_id_array_for_room
+    return Item.where({"room_id" => self.id}).pluck(:id)
+  end
 
   # Method searches all Items for objects associated with the Room, 
   # orders those Items by their "condition" value,
   # and defines the "bad condition" parameter with an array of integers.
   # Items whose condition are included in the array are pushed into an
   # array ("@bad_items").
-  #
-  # RETURNS an Array of Item objects, if not empty (an empty array will return # a nil value).  
-  def find_damaged_items_for_room
-    items = Item.where({"room_id" => self.id}) == nil ? nil : Item.where({"room_id" => self.id}).order('condition')
-    bad_condition = [1, 2, 3]
-    @bad_items = []
-
-    if items != nil
-      items.each do |item|
-        item_condition = item.condition
-        if bad_condition.include?(item_condition)
-          @bad_items << item
-        end
-      end
-    end 
-    return @bad_items.empty? ? nil : @bad_items
+  def find_find_damaged_items_for_room
+    return Item.where({"room_id" => self.id, "condition" => 1..3}).order('condition')
   end
 
   # Method searches all Items for objects associated with the Room, 
@@ -33,21 +30,8 @@ class Room < ActiveRecord::Base
   #
   # RETURNS an Array of Item objects, if not empty (an empty array will return # a nil value).  
   def find_good_items_for_room
-    items = Item.where({"room_id" => self.id}) == nil ? nil : Item.where({"room_id" => self.id}).order('condition desc')
-    good_condition = [4, 5]
-    @good_items = []
-
-    if items != nil
-      items.each do |item|
-        item_condition = item.condition
-        if good_condition.include?(item_condition)
-          @good_items << item
-        end
-      end
-    end 
-    return @good_items.empty? ? nil : @good_items
+    return Item.where({"room_id" => self.id, "condition" => 4..5}).order('condition desc')
   end
-
   # Method searches all Items for objects associated with the Room,
   # pushes them each into an array, then searches all Photos for 
   # objects associated with each Item object. Method then deletes all
@@ -55,21 +39,13 @@ class Room < ActiveRecord::Base
   #
   # RETURNS nil
   def find_and_delete_item_photos_for_room
-    if Item.where({"room_id" => self.id}) != nil
-      item_arr = []
-      items = Item.where({"room_id" => self.id})
-
-      items.each do |item|
-        item_arr << item.id
-      end
-
-      item_arr. each do |item|
-        if Photo.where({"item_id" => item}) != nil
-          Photo.where({"item_id" => item}).delete_all
+    item_id_array = find_items_id_array_for_room
+      item_id_array. each do |item|
+        item_photo = Photo.where({"item_id" => item}) 
+        if item_photo != nil
+          item_photo.delete_all
         end
       end
-
-    end
   end
 
   # Method searches all Items for objects associated with the Room,
@@ -77,35 +53,39 @@ class Room < ActiveRecord::Base
   #
   # RETURNS nil
   def find_and_delete_items_for_room
-    if Item.where({"room_id" => self.id}) != nil
-      Item.where({"room_id" => self.id}).delete_all
-    end
+     Item.where({"room_id" => self.id}).delete_all if @items_collection != nil
   end
 
   # Method RETURNS the User object associated with the "created_by"
   # attribute of the Room.
   #
   # RETURNS a User object, or nil if "created_by" has no value
+  def set_created_by_user_info_for_room
+    @creator = User.find_by_id(self.created_by)
+  end
+
   def get_created_by_user_info_for_room
-    return self.created_by == nil ?  nil : User.find_by_id(self.created_by)
+    return @creator
   end
 
   # Method RETURNS the User object associated with the "updated_by"
   # attribute of the Room.
   #
   # RETURNS a User object, or nil if "updated_by" has no value
-  def get_updated_by_user_info_for_room
-    return self.updated_by == nil ?  nil : User.find_by_id(self.updated_by)
+  def find_updated_by_user_info_for_room
+    @editor = User.find_by_id(self.updated_by)
   end
 
-  # RETURNS Boolean if title value String is empty or not
-  def check_create_room_title_is_valid
-      return self.title != "" ? true : false
+  def get_updated_by_user_info_for_room
+    return @editor
   end
 
   # RETURNS Boolean if location value String is empty or not
   def check_create_room_location_is_valid
-      return self.location != nil ? true : false
+    location = self.location != nil ? true : false
+    if location == false
+      @error << "Please set the location of your room."
+    end
   end
 
   # Method calls on pre-existing Methods to validate input data
@@ -113,30 +93,17 @@ class Room < ActiveRecord::Base
   #
   # RETURNS an Array of errors, which are Strings
   # (created by Methods "room_title_error" and "room_location_error") 
-  def create_room_check_valid_action
-    title = check_create_room_title_is_valid
-    location = check_create_room_location_is_valid
-    message = []
-
+  def check_create_room_title_is_valid
+    title = self.title != "" ? true : false
     if title == false
-      message << room_title_error
+      @error << "Please include a room type."
     end
+  end
 
-    if location == false
-      message << room_location_error
-    end
-
-    return message
+  def error_check
+    check_create_room_title_is_valid
+    check_create_room_location_is_valid
   end
   
-  # RETURNS an error in the form of a String
-  def room_title_error
-    return "Please include a room type."
-  end
-
-  # RETURNS an error in the form of a String
-  def room_location_error
-    return "Please set the location of your room."
-  end
   
 end
